+++ R/chat.R
@@ -16,9 +16,14 @@ chat_deps <- function() {
     src = "lib/shiny",
     script = list(
       list(src = "chat/chat.js", type = "module"),
+      list(src = "markdown-stream/markdown-stream.js", type = "module"),
       list(src = "text-area/textarea-autoresize.js", type = "module")
     ),
-    stylesheet = c("chat/chat.css", "text-area/textarea-autoresize.css")
+    stylesheet = c(
+      "chat/chat.css", 
+      "markdown-stream/markdown-stream.css",
+      "text-area/textarea-autoresize.css"
+    )
   )
 }