+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinychat (development version)
 
+* Added new `output_markdown_stream()` and `markdown_stream()` functions to allow for streaming markdown content to the client. This is useful for showing Generative AI responses in real-time in a Shiny app, outside of a chat interface. (#23)
+
 # shinychat 0.1.1
 
 * Initial CRAN submission.